+++ src/components/common/buttons/oblong-button.tsx
@@ -2,8 +2,8 @@ import { Button, ButtonProps } from '@/components/common/buttons/button';
 
 type OblongButtonProps = Omit<ButtonProps, 'btype'>;
 
-export function OblongButton({ label, link, size }: OblongButtonProps) {
+export function OblongButton({ label, link, size, bgcolor }: OblongButtonProps) {
     return (
-        <Button label={label} link={link} size={size} btype="oblong" />
+        <Button label={label} link={link} size={size} bgcolor={bgcolor} btype="oblong" />
     );
 }