+++ src/components/common/buttons/rectangular-button.tsx
@@ -2,8 +2,8 @@ import { Button, ButtonProps } from '@/components/common/buttons/button';
 
 type RectangularButtonProps = Omit<ButtonProps, 'btype'>;
 
-export function RectangularButton({ label, link, size }: RectangularButtonProps) {
+export function RectangularButton({ label, link, size, bgcolor }: RectangularButtonProps) {
     return (
-        <Button label={label} link={link} size={size} btype="rectangular" />
+        <Button label={label} link={link} size={size} bgcolor={bgcolor} btype="rectangular" />
     );
 }